requirements.txt
Flask
pymongo 
python-dotenv
werkzeug
flask_jwt_extended
PyJWT==2.9.0
pydantic
pydantic[email]